Modifica o id do client para tipo long

+++ src/main/java/com/aguaja/api/domain/Client.java
@@ -21,7 +21,7 @@ public class Client implements Serializable{
 
 	@Id
 	@GeneratedValue(strategy = GenerationType.IDENTITY)
-	private Integer id;
+	private Long id;
 	private String name;
 	private Integer gender;
 	private String email;
@@ -44,7 +44,7 @@ public class Client implements Serializable{
 	public Client() {
 	}
 
-	public Client(Integer id, String name, Integer gender, String email, Instant birthDate, String login, String password, Boolean admin) {
+	public Client(Long id, String name, Integer gender, String email, Instant birthDate, String login, String password, Boolean admin) {
 		this.id = id;
 		this.name = name;
 		this.gender = gender;
@@ -55,10 +55,10 @@ public class Client implements Serializable{
 		this.admin = admin;
 	}
 
-	public Integer getId() {
+	public Long getId() {
 		return id;
 	}
-	public void setId(Integer id) {
+	public void setId(Long id) {
 		this.id = id;
 	}
